+++ components/engine/engine-listeners/src/main/java/org/eclipse/dirigible/components/listeners/service/MessageConsumer.java
@@ -1,3 +1,13 @@
+/*
+ * Copyright (c) 2023 SAP SE or an SAP affiliate company and Eclipse Dirigible contributors
+ *
+ * All rights reserved. This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-FileCopyrightText: 2023 SAP SE or an SAP affiliate company and Eclipse Dirigible
+ * contributors SPDX-License-Identifier: EPL-2.0
+ */
 package org.eclipse.dirigible.components.listeners.service;
 
 import javax.jms.Destination;
@@ -24,15 +34,15 @@ public class MessageConsumer {
         this.session = session;
     }
 
-    public String receiveMessageFromQueue(String queue, int timeout) throws JMSException, TimeoutException {
+    public String receiveMessageFromQueue(String queue, long timeout) throws JMSException, TimeoutException {
         return receiveMessage(timeout, session.createQueue(queue));
     }
 
-    public String receiveMessageFromTopic(String topic, int timeout) throws JMSException, TimeoutException {
+    public String receiveMessageFromTopic(String topic, long timeout) throws JMSException, TimeoutException {
         return receiveMessage(timeout, session.createTopic(topic));
     }
 
-    private String receiveMessage(int timeout, Destination destination) throws JMSException, TimeoutException {
+    private String receiveMessage(long timeout, Destination destination) throws JMSException, TimeoutException {
         try (javax.jms.MessageConsumer consumer = session.createConsumer(destination)) {
 
             Message message = consumer.receive(timeout);
